Add button to allow user to set grid size
Clicking the button will prompt the user to enter the number of
squares, per side, for the new grid. This will be available on page
load, after the default grid (16x16) has already been created via a
newly added page load event (DOMContentLoaded) that ensures the default
grid (16x16) is created and available to the user when arriving on
the site.

If the user chooses, he can create a new grid or use the default grid.
Regardless of the user's choice, the following functionality will be
available:

As the user moves their mouse over a div (mouseenter), it's background
color will change. When leaving the div's boundaries (mouseleave), it's
background color will remain the same. This will appear as a pixelated
trail through the grid like a pen would on an 'Etch-A-Sketch'.